pyproject.toml
[project]
name = "clientadvisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.1.0",
    "jinja2>=3.1.6",
    "langchain[anthropic]>=1.2.11",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langgraph>=1.1.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "pypdf>=6.8.0",
    "sentence-transformers>=5.2.3",
    "langchain-tavily>=0.2.17",
    "ddgs>=9.11.3",
]
